+++ lib/src/time/calendar/calendars/ct_state_holiday_calendar.dart
@@ -1,18 +1,11 @@
 import 'package:date/date.dart';
 import 'package:elec/src/time/calendar/calendar.dart';
 import 'package:elec/src/time/calendar/holiday.dart';
-import '../holidays/lincoln_birthday.dart';
-import '../holidays/good_friday.dart';
-import 'federal_holidays_calendar.dart';
 
 
 /// Connecticut has Lincoln's Birthday, Good Friday in addition to the
 /// Federal holidays.
 class CtStateHolidayCalendar extends Calendar {
-  static final _federalCalendar = FederalHolidaysCalendar();
-  static final Holiday _lincolnBirthday = LincolnBirthday();
-  static final Holiday _goodFriday = GoodFriday();
-
   late HolidayType _holidayType;
 
   @override
@@ -26,32 +19,25 @@ class CtStateHolidayCalendar extends Calendar {
   @override
   bool isHoliday(Date date) {
     var res = false;
-    if (_federalCalendar.isHoliday(date)) {
-      _holidayType = _federalCalendar.getHolidayType(date);
+    if (Calendar.federalHolidays.isHoliday(date)) {
+      _holidayType = Calendar.federalHolidays.getHolidayType(date);
       return true;
     }
 
     if (date.month == 2) {
-      if (_lincolnBirthday.isDate(date)) {
+      if (Holiday.lincolnBirthday.isDate(date)) {
         res = true;
         _holidayType = HolidayType.lincolnBirthday;
       }
 
     } else if (date.month == 3 || date.month == 4) {
-      if (_goodFriday.isDate(date)) {
+      if (Holiday.goodFriday.isDate(date)) {
         res = true;
         _holidayType = HolidayType.goodFriday;
       }
     }
     return res;
   }
-
-  @override
-  bool isHoliday3(int year, int month, int day) {
-    // TODO: implement isHoliday3
-    throw UnimplementedError();
-  }
-
 }