+++ lib/src/time/calendar/calendars/ma_state_holiday_calendar.dart
@@ -1,16 +1,11 @@
 import 'package:date/date.dart';
 import 'package:elec/src/time/calendar/calendar.dart';
 import 'package:elec/src/time/calendar/holiday.dart';
-import '../holidays/patriots_day.dart';
-import 'federal_holidays_calendar.dart';
 
 
 /// Massachusetts has Patriots' Day in addition to the
 /// 10 Federal holidays.
 class MaStateHolidayCalendar extends Calendar {
-  static final _federalCalendar = FederalHolidaysCalendar();
-  static final Holiday _patriotsDay = PatriotsDay();
-
   late HolidayType _holidayType;
 
   @override
@@ -24,26 +19,19 @@ class MaStateHolidayCalendar extends Calendar {
   @override
   bool isHoliday(Date date) {
     var res = false;
-    if (_federalCalendar.isHoliday(date)) {
-      _holidayType = _federalCalendar.getHolidayType(date);
+    if (Calendar.federalHolidays.isHoliday(date)) {
+      _holidayType = Calendar.federalHolidays.getHolidayType(date);
       return true;
     }
 
     if (date.month == 4) {
-      if (_patriotsDay.isDate(date)) {
+      if (Holiday.patriotsDay.isDate(date)) {
         res = true;
         _holidayType = HolidayType.patriotsDay;
       }
     }
     return res;
   }
-
-  @override
-  bool isHoliday3(int year, int month, int day) {
-    // TODO: implement isHoliday3
-    throw UnimplementedError();
-  }
-
 }