+++ lib/src/time/calendar/calendars/nh_state_holiday_calendar.dart
@@ -37,15 +37,6 @@ class NhStateHolidayCalendar extends Calendar {
     }
     return res;
   }
-
-  @override
-  bool isHoliday3(int year, int month, int day) {
-    // TODO: implement isHoliday3
-    throw UnimplementedError();
-  }
-
-
-
 }